🧑‍💻 (Dotenv and requirements) Add .env.example file and un-freeze requirements.txt
Since we're using `DATABASE_URL` in the app, added `.env.example` file to tell other devs that they should create a `.env` file to run the project locally.

Also un-froze the `requirements.txt` file, only adding what the project needs.

Added `python-dotenv` as a dependency so that when we execute the app using `flask run`, it will load the environment variables in `.env` automatically (which should include `DATABASE_URL`).