🗑️ (Remove `app.run()`) Use the Flask CLI instead of `app.run()` in `app.py`
As of Flask v1.0, we no longer need to use the `if __name__ == "__main__"` statement, or do `app.run()`. Instead, we can use the Flask CLI (comes with `pip install flask`) and execute the `flask run` command. This also has the benefit of loading environment variables before starting execution (if the `python-dotenv` package is installed).

As part of this commit I also removed a duplicate `.env` entry from `.gitignore`, and removed `.flaskenv` too.